+++ tools/transformer/src/atlasTransformations.js
@@ -10,6 +10,10 @@ const {
   removeRefsFromParameters,
 } = require("./transformations");
 
+const {
+  resolveOpenAPIReference
+} = require("./engine/transformers");
+
 const removeUnusedSchemas = require("./engine/removeUnused");
 
 const ignoredModelNames = require("./name.ignore.json").ignoreModels;
@@ -74,41 +78,25 @@ module.exports = function runTransformations(openapi) {
 function searchAPIIssuesTransformation(openapi) {
   // API is overly complex in this case and provides no value as typed interface
   if (openapi.components.schemas.SearchIndexResponse) {
-    responseParent = openapi.components.schemas.SearchIndexResponse;
-    if (responseParent.properties.latestDefinition) {
-      if(responseParent.discriminator.mapping){
-        responseParent.properties.latestDefinition = {oneOf: []};
-        for(mappingKey in responseParent.discriminator.mapping){
-          var ref = responseParent.discriminator.mapping[mappingKey];
-          var reference = resolveOpenAPIReference(openapi,ref)
+    const responseParent = openapi.components.schemas.SearchIndexResponse;
+    if (responseParent.properties && responseParent.properties.latestDefinition) {
+      if (responseParent.discriminator && responseParent.discriminator.mapping) {
+        responseParent.properties.latestDefinition = { oneOf: [] };
+        for (const mappingKey in responseParent.discriminator.mapping) {
+          const ref = responseParent.discriminator.mapping[mappingKey];
+          if (!ref) {
+            continue; // Skip if there's no reference
+          }
+          const reference = resolveOpenAPIReference(openapi, ref);
           responseParent.properties.latestDefinition.oneOf.push({
-            $ref: reference.allOf[1].properties.latestDefinition.$ref
+            $ref: reference.allOf && reference.allOf[1] && reference.allOf[1].properties && reference.allOf[1].properties.latestDefinition && reference.allOf[1].properties.latestDefinition.$ref
           });
-          delete reference.allOf[1].properties.latestDefinition
+          delete reference.allOf[1]?.properties?.latestDefinition;
         }
       }
     }
   }
-  console.error("SearchIndexResponse", openapi.components.schemas.SearchIndexResponse)
-  console.error("VectorSearchIndexResponse", openapi.components.schemas.VectorSearchIndexResponse)
-  console.error("TextSearchIndexResponse", openapi.components.schemas.TextSearchIndexResponse)
+  console.debug("SearchIndexResponse", openapi.components?.schemas?.SearchIndexResponse);
   return openapi;
 }
 
-function resolveOpenAPIReference(openapi, ref) {
-  if (!ref.startsWith("#/")) {
-    throw new Error("Invalid reference format: " + ref);
-  }
-  const parts = ref.split("/");
-  // Skip the first empty part
-  parts.shift();
-
-  let current = openapi;
-  for (const part of parts) {
-    if (!current.hasOwnProperty(part)) {
-      throw new Error("Reference not found: " + ref);
-    }
-    current = current[part];
-  }
-  return current;
-}
